requirements.txt
python-dotenv>=1.0.0
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
requests>=2.31.0
stripe>=7.0.0
plotly>=5.18.0
